# docs/swagger.yaml
basePath: /
definitions:
...
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  entities.CreateEmpleadoRequest:
    properties:
      apellidos:
        type: string
      desactivado:
        type: boolean
      mail:
        type: string
      nombre:
        type: string
      password:
        type: string
      rol_id:
        type: integer
      username:
        type: string
    required:
    - apellidos
    - mail
    - nombre
    - password
    - username
    type: object
...
  entities.Empleado:
    properties:
      apellidos:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      desactivado:
        type: boolean
      id:
        type: integer
      mail:
        type: string
      nombre:
        type: string
      rol_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  entities.EmpleadoListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.Empleado'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.EmpleadoMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.EmpleadoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Empleado'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.UpdateEmpleadoRequest:
    properties:
      apellidos:
        type: string
      desactivado:
        type: boolean
      mail:
        type: string
      nombre:
        type: string
      password:
        type: string
      rol_id:
        type: integer
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/empleados/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EmpleadoListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar empleados
      tags:
      - Empleado
    post:
      consumes:
      - application/json
      parameters:
      - description: Empleado
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateEmpleadoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.EmpleadoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear empleado
      tags:
      - Empleado
  /api/empleados/{id}:
    delete:
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EmpleadoMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar empleado
      tags:
      - Empleado
    get:
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EmpleadoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener empleado por ID
      tags:
      - Empleado
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateEmpleadoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EmpleadoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar empleado
      tags:
      - Empleado
...
securityDefinitions:
  BearerAuth:
    description: 'Escribe ''Bearer '' seguido de tu token JWT. Ejemplo: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"